requirements.txt
google-auth==2.10.0
google-auth-oauthlib==0.5.2
google-auth-httplib2==0.1.0
google-api-python-client==2.52.0
requests==2.28.1
